Factor out a lot of logic
