+++ jest.config.js
@@ -5,6 +5,7 @@ module.exports = {
     '\\.css$': '<rootDir>/__mocks__/styleMock.js'
   },
   transformIgnorePatterns: ['<rootDir>/node_modules/(?!(base2048|base65536)/)'],
+  maxWorkers: 1,
   coverageThreshold: {
     global: {
       branches: 100,
